supporting tensorboard's scalars by batch step

+++ utils.py
@@ -10,10 +10,14 @@ from data import ThumbnailsDataset
 import matplotlib.pyplot as plt
 import sklearn.model_selection
 from PIL.Image import Image
+import pathlib
+
+BATCH_SIZE = 8
 
 
 def REAL_PATH(path):
-    return os.path.join('/home/haneenna/GipMed-project-234329', path)
+    print(os.path.join(os.path.abspath(os.getcwd()), path))
+    return os.path.join(os.path.abspath(os.getcwd()), path)
 
 def save_checkpoint(state, filename="my_checkpoint.pth.tar"):
     print("=> Saving checkpoint")
@@ -26,23 +30,31 @@ def load_checkpoint(model, filename="my_checkpoint.pth.tar"):
     model.load_state_dict(checkpoint["model"])
 
 
-def check_accuracy(loader, model, device="cuda"):
+def check_accuracy(loader, model ,epoch ,writer ,device="cuda"):
     num_correct = 0
     num_pixels = 0
     dice_score = 0
     model.eval()  # set the model to be in eval mode not train mode, for parts that behave differently in train/val
 
     with torch.no_grad():
-        for x, y in loader:
+        for cur_batch, (x, y) in enumerate(loader):
             x = x.to(device)
             y = y.to(device).unsqueeze(1)
             preds = torch.sigmoid(model(x))
             preds = (preds > 0.5).float()
-            num_correct += (preds == y).sum()
-            num_pixels += torch.numel(preds)
-            dice_score += (2 * (preds * y).sum()) / (
+            cur_correct = (preds == y).sum()
+            num_correct += cur_correct
+            cur_pixels = torch.numel(preds)
+            num_pixels += cur_pixels
+            cur_dice = (2 * (preds * y).sum()) / (
                     (preds + y).sum() + 1e-8
             )
+            dice_score += cur_dice
+            global_batch_counter = BATCH_SIZE*epoch + cur_batch
+            acc_per_batch = float(cur_correct/cur_pixels * 100)
+            dice_per_batch = float(cur_dice/BATCH_SIZE)
+            writer.add_scalar('ACC/validation', acc_per_batch, global_batch_counter)
+            writer.add_scalar('DICE_SCORE/validation', dice_per_batch, global_batch_counter)
 
     print(
         f"Got {num_correct}/{num_pixels} with acc {num_correct / num_pixels * 100:.2f}"
@@ -150,7 +162,10 @@ def extractingPhotosProperties(img_dirs):
 
 
 def clear_folder(dir):
-    files = os.listdir(REAL_PATH(dir))
-    for f in files:
-        f_path = os.path.join(dir, f)
-        os.remove(f_path)
+    if os.path.exists(REAL_PATH(dir)):
+        files = os.listdir(REAL_PATH(dir))
+        for f in files:
+            f_path = os.path.join(dir, f)
+            os.remove(f_path)
+    else:
+        os.mkdir(REAL_PATH(dir))
